Move update node functions out of dataflow-program.tsx.
